requirements.txt
streamlit>=1.36.0
plotly>=5.18.0
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
lxml>=5.0.0
py7zr>=1.0.0
